# internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/admin/invoke/count": {
            "post": {
                "description": "get api invoke count",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "invoke"
                ],
                "summary": "get api invoke count",
                "parameters": [
                    {
                        "description": "invoke request",
                        "name": "invokeRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.Invoke"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    }
                }
            }
        },
        "/admin/question/create_question": {
            "post": {
                "description": "Create question",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "question"
                ],
                "summary": "Create question",
                "parameters": [
                    {
                        "description": "create question",
                        "name": "createRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.Question"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    }
                }
            }
        },
        "/admin/question/delete_question": {
            "get": {
                "description": "Delete question",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "question"
                ],
                "summary": "Delete question",
                "parameters": [
                    {
                        "type": "string",
                        "description": "delete question",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    }
                }
            }
        },
        "/admin/question/update_question": {
            "post": {
                "description": "Update question",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "question"
                ],
                "summary": "Update question",
                "parameters": [
                    {
                        "description": "update question",
                        "name": "updateRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateQuestion"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    }
                }
            }
        },
...
        "/user/picture/avatar": {
            "get": {
                "description": "Get random picture",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "picture"
                ],
                "summary": "Get random picture",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    }
                }
            }
        },
        "/user/question/get_questions": {
            "get": {
                "description": "get question list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "question"
                ],
                "summary": "get question list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "page number",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "page size",
                        "name": "pageSize",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    }
                }
            }
        },
        "/user/question/search_questions": {
            "get": {
                "description": "Search question",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "question"
                ],
                "summary": "Search question",
                "parameters": [
                    {
                        "type": "string",
                        "description": "search question",
                        "name": "keyword",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "common.Config": {
            "type": "object"
        },
        "middlewares.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "request.Invoke": {
            "type": "object",
            "required": [
                "method",
                "path"
            ],
            "properties": {
                "method": {
                    "type": "string",
                    "enum": [
                        "GET",
                        "POST"
                    ]
                },
                "path": {
                    "type": "string"
                }
            }
        },
...
        "request.Question": {
            "type": "object"
        },
...
        "request.UpdateQuestion": {
            "type": "object",
            "required": [
                "id",
                "user_id"
            ],
            "properties": {
                "accept_num": {
                    "description": "\"题目通过数\"",
                    "type": "integer"
                },
                "answer": {
                    "description": "\"题目答案,数组\"",
                    "type": "array",
                    "maxItems": 512,
                    "items": {
                        "type": "string"
                    }
                },
                "content": {
                    "description": "\"内容\"",
                    "type": "string",
                    "maxLength": 1024
                },
                "id": {
                    "type": "integer",
                    "maximum": 10000000000000000000
                },
                "judge_case": {
                    "description": "\"判题用例json数组,输入用例\"",
                    "type": "array",
                    "maxItems": 1024,
                    "items": {
                        "type": "string"
                    }
                },
                "judge_config": {
                    "description": "\"判题配置json对象,内存限制，时间限制\"",
                    "type": "array",
                    "maxItems": 64,
                    "items": {
                        "$ref": "#/definitions/common.Config"
                    }
                },
                "submit_num": {
                    "description": "\"题目提交数",
                    "type": "integer"
                },
                "tag": {
                    "description": "\"标签列表json数组\"",
                    "type": "array",
                    "maxItems": 128,
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "description": "\"标题\"",
                    "type": "string",
                    "maxLength": 512
                },
                "user_id": {
                    "description": "\"创建用户id\"",
                    "type": "integer",
                    "maximum": 64
                }
            }
        },
...
    }
}